Add AsyncThrowingStream.init(unfolding:onCancel:) to handle cancellation.
It shares implementation with and becomes the general case of the existing `init(unfolding:)`, and given the current implementation's use of `withTaskCancellationHandler`, the `onCancel` closure can run concurrently with the `unfolding` closure if cancellation happens during production, or ahead of it when cancellation has already taken place. See doc and the last 3 test cases added.

Also adds tests for `init(unfolding:)`.

Resolves #77974.