[Observation] Add type annotation on cached keypaths
Having a KeyPath with an inferred type that contains a "some" type gives
an error, like:

  error: property definition has inferred type '...', involving the
   'some' return type of another declaration

This is not diagnosed if there is an explicit type, so add an explicit
type for the cached keypath when the property has an annotated type.
This doesn't resolve cases where there is no type annotation on a
property and the inferred type contains a `some` type, but this can
always be resolved by adding an explicit type.

rdar://143466616