[Mangler] Handle ElementArchetypeType for decls
Element archetypes can occur when mangling the USR
for local variables for e.g SourceKit cursor info,
as well as for regular compilation for things like
lazy variables.

This patch adds some basic mangling logic to handle
mangling the type for a decl, which is sufficient
to fix things like cursor info. More work is needed
to support lazy variables though (now it crashes in
SILGen).

This patch doesn't handle mangling standalone
element archetypes for e.g `printTypeUSR`, we'll
either need to fixup each of the clients there, or
we'll need to rethink how we form local archetypes
to correctly handle that.

rdar://143077965