[windows] Enable tests for LLDB in Swift CI (#78844)
TL;DR Please revert this patch in case LLDB testing on Windows clogs Swift PR-tests

The LLDB test suite is not very stable on Windows, which is a problem when we want to run them in CI. In order to improve the situation, however, we need to get it tested continuously. This patch attempts to solve the chicken-egg problem. It introduces a simple override for unstable tests: list them in file `utils/windows-llvm-lit-test-overrides.txt` and SwiftCI will skip/xfail them.

This approach has a number of benefits:
* overrides don't need to be in sync with https://github.com/swiftlang/llvm-project
* overrides are tracked in one place and they are not spread across the LLDB test suite
* overrides are swiftlang-specific, which clearly states the differences to upstream LLVM
* we can enable continuous testing (and get reports for new failures) without fixing the world first

Once the remaining subset of tests passes reliably, we can iterate to enable the others and increase coverage. My previous patch 459e59265d5f99cb18435f1728eff74ffabd4c1f implemented the necessary infrastructure. This one only enable tests in CI. The separation allows to disable LLDB testing on Windows with fewer side-effects.