[CSOptimizer] Prefer some unsupported/un-scored disjunctions over operators
For example: `a.b() + 1` or a more complex `[a.b()].map { $0 + 1 }`.

if `b` has active generic overloads it would be deemed unsupported
but selecting it before `+` (which has a low score due to a literal use)
is preferable because it would either put additional generic requirements
on the argument of `+` or provide a concrete type that would be used
to further filter operator's overload set.

Selecting operator first in these cases could result in the solver
having to attempt every operator overload for every overload of `b`.