[Backtracing] Declare a CMake dependency on CxxStdlib overlay
The Swift code in the Backtracing library enables C++ interop and uses C++ libraries that `#include` C++ stdlib headers. This means that Swift will pull in the C++ stdlib module, along with its overlay. The overlay module needs to be already built at that point.

This should fix an occasional compiler error:
```
error: could not find module 'CxxStdlib' for target
```