Sema: Fix generic type alias resolution edge case
If a type alias in generic context fixes all outer generic
parameters to concrete types, we allow the type alias to
be referenced without specifying the generic arguments of
its parent type.

However, we need to reduce the underlying type in case it
was written in terms of the (fully concrete) generic
parameters.

Fixes rdar://problem/143707820.