build-script-impl: propagate `--verbose-build` to nested CMake builds
The main goal of this change is to ensure that the new build of the
stdlib matches the same level of verbosity of the compiler build that
spawn it.

For now I'm not matching this behaviour to the regular CMake build
products (which would be needed if want to target external projects
configured in LLVM).

Addresses rdar://144256800